+++ src/main/java/io/github/prluciohermano/domain/repository/ItemsPedido.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.github.prluciohermano.domain.entity.ItemPedido;
 
-public interface ItemsPedido extends JpaRepository<ItemPedido, Integer> {
+public interface ItemsPedido extends JpaRepository<ItemPedido, Long> {
 
 }
